+++ packages/fastify/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @vercel/fastify
 
+## 0.1.60
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vercel/node@5.6.24
+
+## 0.1.59
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vercel/node@5.6.23
+
 ## 0.1.58
 
 ### Patch Changes